refactor(k6): organize test structure
