+++ .gitignore
@@ -56,3 +56,4 @@ monitoring/alertmanager.yml
 
 ### K6 data ###
 k6/reports/**
+!k6/reports/.gitkeep
